+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.1.8 - 2017-06-08
+
+### Added
+- lang option in Github.get_starred_repos
+- sample scripts to printout github daily trends
+
 ## 0.1.7 - 2017-05-23
 
 ### Added